requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
Django==5.0.7
pytz==2024.1
six==1.16.0
speedtest-cli==2.1.3
sqlparse==0.5.0
tzdata==2024.1
tzlocal==5.2
